uploading checkin changes plus header updates

+++ src/main/java/com/techelevator/dao/GymCheckinDao.java
@@ -6,7 +6,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 public interface GymCheckinDao {
-    List<GymCheckin> getCheckInLogForUser(int userId);
+    List<GymCheckin> getCheckInLogForUser(long userId);
     List<GymCheckin> getAllUserCheckInLog();
 
 }
